update test modules
