+++ src/candidate.rs
@@ -488,147 +488,13 @@ impl Candidate {
 
 #[cfg(test)]
 mod test {
-    use crate::prelude::{
-        Candidate, Carrier, Duration, Epoch, PhaseCombination, PhaseRange, PseudoRange,
-        PseudoRangeCombination, SV,
-    };
-    #[test]
-    fn prefered_pseudorange() {
-        for (observations, prefered) in [
-            (
-                vec![
-                    Observation {
-                        value: 1.0,
-                        snr: None,
-                        carrier: Carrier::L1,
-                    },
-                    Observation {
-                        value: 2.0,
-                        snr: None,
-                        carrier: Carrier::L2,
-                    },
-                    Observation {
-                        value: 3.0,
-                        snr: None,
-                        carrier: Carrier::L5,
-                    },
-                ],
-                Observation {
-                    value: 1.0,
-                    snr: None,
-                    carrier: Carrier::L1,
-                },
-            ),
-            (
-                vec![
-                    Observation {
-                        value: 1.0,
-                        snr: None,
-                        carrier: Carrier::L1,
-                    },
-                    Observation {
-                        value: 2.0,
-                        snr: Some(2.0),
-                        carrier: Carrier::L2,
-                    },
-                    Observation {
-                        value: 3.0,
-                        snr: None,
-                        carrier: Carrier::L5,
-                    },
-                ],
-                Observation {
-                    value: 2.0,
-                    snr: Some(2.0),
-                    carrier: Carrier::L2,
-                },
-            ),
-        ] {
-            let cd = Candidate::new(
-                SV::default(),
-                Epoch::default(),
-                Duration::default(),
-                None,
-                observations,
-                vec![],
-                vec![],
-            );
-            assert_eq!(cd.prefered_pseudorange(), Some(prefered),);
-        }
-    }
-    #[test]
-    fn best_snr() {
-        let codes = vec![
-            Observation {
-                value: 1.0,
-                snr: None,
-                carrier: Carrier::L1,
-            },
-            Observation {
-                value: 2.0,
-                snr: None,
-                carrier: Carrier::L2,
-            },
-            Observation {
-                value: 3.0,
-                snr: Some(10.0),
-                carrier: Carrier::L5,
-            },
-            Observation {
-                value: 4.0,
-                snr: Some(11.0),
-                carrier: Carrier::L2,
-            },
-            Observation {
-                value: 5.0,
-                snr: Some(9.0),
-                carrier: Carrier::L2,
-            },
-        ];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            codes,
-            vec![],
-            vec![],
-        );
-        assert_eq!(cd.best_snr(), Some(11.0));
-
-        let codes = vec![
-            Observation {
-                value: 1.0,
-                snr: Some(1.0),
-                carrier: Carrier::L1,
-            },
-            Observation {
-                value: 2.0,
-                snr: Some(1.1),
-                carrier: Carrier::L2,
-            },
-            Observation {
-                value: 3.0,
-                snr: Some(1.2),
-                carrier: Carrier::L5,
-            },
-        ];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            codes,
-            vec![],
-            vec![],
-        );
-        assert_eq!(cd.best_snr(), Some(1.2));
-    }
+    use super::{PhaseCombination, PseudoRangeCombination};
+    use crate::prelude::{Candidate, Carrier, Duration, Epoch, PhaseRange, PseudoRange, SV};
     #[test]
     fn cpp_compatibility() {
         for (pr_observations, phase_observations, cpp_compatible) in [
             (
-                vec![Observation {
+                vec![PseudoRange {
                     value: 1.0,
                     snr: Some(1.0),
                     carrier: Carrier::L1,
@@ -638,12 +504,12 @@ mod test {
             ),
             (
                 vec![
-                    Observation {
+                    PseudoRange {
                         value: 1.0,
                         snr: Some(1.0),
                         carrier: Carrier::L1,
                     },
-                    Observation {
+                    PseudoRange {
                         value: 2.0,
                         snr: Some(2.0),
                         carrier: Carrier::L2,
@@ -660,273 +526,8 @@ mod test {
                 None,
                 pr_observations,
                 phase_observations,
-                vec![],
             );
             assert_eq!(cd.cpp_compatible(), cpp_compatible);
         }
-    }
-    #[test]
-    fn l1_phase_widelane() {
-        let phases = vec![
-            Observation {
-                snr: None,
-                value: 64.0,
-                carrier: Carrier::L1,
-            },
-            Observation {
-                snr: None,
-                value: 128.0,
-                carrier: Carrier::L2,
-            },
-        ];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            vec![],
-            phases,
-            vec![],
-        );
-        let pw = cd.phase_wl_combination();
-        assert!(pw.is_some(), "failed to form Ph_wide combination");
-        let pw = pw.unwrap();
-        assert_eq!(pw.carrier, Carrier::L1);
-        assert_eq!(
-            pw.value,
-            (Carrier::L1.frequency() * 64.0 - Carrier::L2.frequency() * 128.0)
-                / (Carrier::L1.frequency() - Carrier::L2.frequency())
-        );
-
-        let phases = vec![Observation {
-            snr: None,
-            value: 64.0,
-            carrier: Carrier::L1,
-        }];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            vec![],
-            phases,
-            vec![],
-        );
-
-        assert!(
-            cd.phase_wl_combination().is_none(),
-            "Ph_wide should not be feasible!"
-        );
-    }
-    #[test]
-    fn e1_phase_widelane() {
-        let phases = vec![
-            Observation {
-                snr: None,
-                value: 64.0,
-                carrier: Carrier::E1,
-            },
-            Observation {
-                snr: None,
-                value: 128.0,
-                carrier: Carrier::E5,
-            },
-        ];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            vec![],
-            phases,
-            vec![],
-        );
-        let pw = cd.phase_wl_combination();
-        assert!(pw.is_some(), "failed to form Ph_wide combination");
-        let pw = pw.unwrap();
-        assert_eq!(pw.carrier, Carrier::E1);
-        assert_eq!(
-            pw.value,
-            (Carrier::E1.frequency() * 64.0 - Carrier::E5.frequency() * 128.0)
-                / (Carrier::E1.frequency() - Carrier::E5.frequency())
-        );
-    }
-    #[test]
-    fn l1_code_narrowlane() {
-        let codes = vec![
-            Observation {
-                snr: None,
-                value: 64.0,
-                carrier: Carrier::L1,
-            },
-            Observation {
-                snr: None,
-                value: 128.0,
-                carrier: Carrier::L2,
-            },
-        ];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            codes,
-            vec![],
-            vec![],
-        );
-        let cn = cd.code_nl_combination();
-        assert!(cn.is_some(), "failed to form Cn_narrow combination");
-        let cn = cn.unwrap();
-        assert_eq!(cn.carrier, Carrier::L1);
-        assert_eq!(
-            cn.value,
-            (Carrier::L1.frequency() * 64.0 + Carrier::L2.frequency() * 128.0)
-                / (Carrier::L1.frequency() + Carrier::L2.frequency())
-        );
-
-        let codes = vec![Observation {
-            snr: None,
-            value: 64.0,
-            carrier: Carrier::L1,
-        }];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            codes,
-            vec![],
-            vec![],
-        );
-
-        assert!(
-            cd.code_nl_combination().is_none(),
-            "Cn_narrow should not be feasible!"
-        );
-    }
-    #[test]
-    fn e1_code_narrow() {
-        let codes = vec![
-            Observation {
-                snr: None,
-                value: 64.0,
-                carrier: Carrier::E1,
-            },
-            Observation {
-                snr: None,
-                value: 128.0,
-                carrier: Carrier::E5,
-            },
-        ];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            codes,
-            vec![],
-            vec![],
-        );
-        let cn = cd.code_nl_combination();
-        assert!(cn.is_some(), "failed to form Cn_narrow combination");
-        let cn = cn.unwrap();
-        assert_eq!(cn.carrier, Carrier::E1);
-        assert_eq!(
-            cn.value,
-            (Carrier::E1.frequency() * 64.0 + Carrier::E5.frequency() * 128.0)
-                / (Carrier::E1.frequency() + Carrier::E5.frequency())
-        );
-    }
-    #[test]
-    fn e1_e5_mw_combination() {
-        let codes = vec![
-            Observation {
-                snr: None,
-                value: 64.0,
-                carrier: Carrier::E1,
-            },
-            Observation {
-                snr: None,
-                value: 128.0,
-                carrier: Carrier::E5,
-            },
-        ];
-        let phases = vec![
-            Observation {
-                snr: None,
-                value: 16.0,
-                carrier: Carrier::E1,
-            },
-            Observation {
-                snr: None,
-                value: 32.0,
-                carrier: Carrier::E5,
-            },
-        ];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            codes,
-            phases,
-            vec![],
-        );
-        let mw = cd.mw_combination();
-        assert!(mw.is_some(), "failed to form MW combination");
-        let mw = mw.unwrap();
-        assert_eq!(mw.lhs, Carrier::E5);
-        assert_eq!(mw.reference, Carrier::E1);
-        let pw = (Carrier::E1.frequency() * 16.0 - Carrier::E5.frequency() * 32.0)
-            / (Carrier::E1.frequency() - Carrier::E5.frequency());
-        let cn = (Carrier::E1.frequency() * 64.0 + Carrier::E5.frequency() * 128.0)
-            / (Carrier::E1.frequency() + Carrier::E5.frequency());
-        assert_eq!(mw.value, pw - cn);
-    }
-    #[test]
-    fn l1_l5_mw_combination() {
-        let codes = vec![
-            Observation {
-                snr: None,
-                value: 64.0,
-                carrier: Carrier::L1,
-            },
-            Observation {
-                snr: None,
-                value: 128.0,
-                carrier: Carrier::L5,
-            },
-        ];
-        let phases = vec![
-            Observation {
-                snr: None,
-                value: 16.0,
-                carrier: Carrier::L1,
-            },
-            Observation {
-                snr: None,
-                value: 32.0,
-                carrier: Carrier::L5,
-            },
-        ];
-        let cd = Candidate::new(
-            SV::default(),
-            Epoch::default(),
-            Duration::default(),
-            None,
-            codes,
-            phases,
-            vec![],
-        );
-        let mw = cd.mw_combination();
-        assert!(mw.is_some(), "failed to form MW combination");
-        let mw = mw.unwrap();
-        assert_eq!(mw.lhs, Carrier::L5);
-        assert_eq!(mw.reference, Carrier::L1);
-        let pw = (Carrier::L1.frequency() * 16.0 - Carrier::L5.frequency() * 32.0)
-            / (Carrier::L1.frequency() - Carrier::L5.frequency());
-        let cn = (Carrier::L1.frequency() * 64.0 + Carrier::L5.frequency() * 128.0)
-            / (Carrier::L1.frequency() + Carrier::L5.frequency());
-        assert_eq!(mw.value, pw - cn);
     }
 }
