+++ src/tests/mod.rs
@@ -1,6 +1,8 @@
 use crate::prelude::*;
 
-pub mod data;
+mod bancroft;
+mod data;
+mod pseudo_range;
 
 //pub mod cpp;
 //pub mod ppp;
